Split MixerTrackCluster into subclasses
This cleans up a lot of the code that would be shared between note and wave tracks but also allows putting the note track code in its own define.  I did chose to take out a lot of the code that was formerly only used when midi output was disabled, because maintaining two forms is far more complicated than one more general form.

For the most part, the wave code is copied directly; there's a few type changes though to reduce warnings (and additional assertions as to the type of the track).

There are some assertions here that may be invalid; they've been marked as such.