Switch to a dedicated velocity slider in TrackPanel
This gets rid of a large number of ugly hacks regarding the gain slider and offsets, which generally is good.  I've also cleaned up the code regarding the channel labels/toggles on note tracks, so that it doesn't depend on the offset rectangle.

In the future, it may make sense to move the channel label code into TrackPanel; it's odd to have it in NoteTrack.  That would also allow replacement of the constants that were copied into NoteTrack.