Use mTimeTrack timing rather than mMidiPlaySpeed
mMidiPlaySpeed doesn't handle all of the options provided by time tracks,
and in fact can't doesn't even use time tracks.  Using the time track
instead allows midis to remain synchronized with the rest of the audio.

mMidiPlaySpeed _did_ work where it was used in the transcription tool
bar, but that already updates a time track so no special handling is
needed.