Fix channel toggling as EXPERIMENTAL_MIDI_CONTROLS
This reintroduces the buttons to toggle display of individual midi
channels, and cleans up the code behind that feature.  It also adds an
EXPERIMENTAL_MIDI_CONTROLS flag (enabled by default) to use this function,
and puts the related code behind it.

This functionality has actually been present in production versions of
audacity for a while, at least for clicking.  However, the buttons
themselves were not drawn, making it exteremly painful (but possible) to
use.  When EXPERIMENTAL_MIDI_CONTROLS is not enabled, it will no longer be
possible to use these invisible buttons, which means the flag _does_
remove existing functionality, just not functionality that was fully
usable.