Attempt to add the rest of the makefile for portmidi
Also, fix a broken(?) include in px_linux_alsa.c.  I don't know if it was actually broken, but it seems like that was wrong.

The change to portaudio's makefile is definitely incorrect - it shouldn't need manual adding, but I don't know how to choose the right file by OS (or rather, there clearly is code for that, but I don't know how to get it to run).