Fix velocity offset edge case
If the velocity of a note was, for instance, 30, and the velocity slider was at -30, then you'd get a data2 of 0, which wouldn't be clamped (and would produce a note off).  However, if the note were 29 or 31, you'd get a velocity of 1.  This makes it so that note ons always have a velocity of at least 1, even in that edge case.