Get scrubbing to work with midi playback
There are some major issues with this implementation:
 * Synchronization is COMPLETELY disabled; this means that outside of scrubbing desync can happen
 * I'm no longer passing a latency parameter to portmidi.  This means that everything is played as it happens (ie, midi ticks, with a MIDI_SLEEP delay between them).  While the current MIDI_SLEEP (10 ms) is small, it's present.
 * Looping no longer works
 * The issue PRL mentioned of midi not being pitch-scaled is definitely present.  For powers of 2 (.5x, 1x, 2x), this is not that big of a deal, but for anything else, it sounds really bad.
 * Backwards playback still doesn't work.  This will be harder to fix.

The key changes to this were to disable timestamps and just output events as they come (instead of trying to predict user actions), and using mTime for audiotime instead of counting the number of samples output.  Neither of these are perfect, but it's a start, still.

One other thing to note: Seeking also works fine!  I'm not completely sure _why_, but it does... probably something with it running all of the events up to the current time.