Get scrubbing to work with MIDI playback
There are some potential issues with this implementation (which are due to the nature of the change):
 * I'm no longer passing a latency parameter to portmidi during scrubbing.  This means that everything is played as it happens (ie, MIDI ticks, with a MIDI_SLEEP delay between them).  While the current MIDI_SLEEP (10 ms) is small, it's present.
 * The issue PRL mentioned of MIDI not being pitch-scaled is definitely present.  For powers of 2 (.5x, 1x, 2x), this is not that big of a deal, but for anything else, it sounds really bad.
 * Backwards playback still doesn't work.  This will be harder to fix.

The key changes to this were to disable timestamps and just output events as they come (instead of trying to predict user actions), and using mTime for AudioTime instead of counting the number of samples output.  Neither of these are perfect, and both of these cause problems if used outside of scrubbing, but for scrubbing they work fine.

One other thing to note: Seeking also works fine!  I'm not completely sure _why_, but it does... probably something with it running all of the events up to the current time.