Bug 1690: Playback does not stop at the end of MIDI playback
This was caused by the indicator-passed-selection check being inside a loop over all wave tracks.  If there are no wave tracks, then the check never happened.  I added a duplicate version of the check if there are no wave tracks.  This may not be the best solution, but it works for now.

I'm not entirely sure why the check is in the loop (or what's going on with the len and maxLen variables); as such, I just haven't touched that.  Perhaps it would be safe to move the check always outside of the loop, but I'm not completely confident in that.